+++ seiscat/utils/__init__.py
@@ -11,4 +11,6 @@ Utility functions for seiscat.
 """
 from .exit import err_exit, ExceptionExit  # noqa
 from .conversion import float_or_none, int_or_none  # noqa
+from .sample_script import write_sample_script  # noqa
 from .print_logo import print_logo  # noqa
+from .write_ok import write_ok  # noqa